removed errors and added logging

+++ Ember/src/Ember/Platform/Linux/LinxWindow.cpp
@@ -103,7 +103,7 @@ namespace Ember {
 		glfwSetCharCallback(m_Window, [](GLFWwindow* window, unsigned int keycode)
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			KeyTypedEvent event(keycode);
+			KeyTypedEvent event((int)keycode);
 			data.EventCallback(event);
 		});
